+++ cppcore/include/system/Foundation.hpp
@@ -24,12 +24,13 @@ namespace detail {
     /// Initialize the neighbor count for each site
     ArrayX<int16_t> count_neighbors(Foundation const& foundation);
     /// Reduce this site's neighbor count to zero and inform its neighbors of the change
-    void clear_neighbors(Site& site, ArrayX<int16_t>& neighbor_count);
-    /// Remove edge sites which have a neighbor count lower than the lattice minimum
-    void trim_edges(Foundation& foundation);
+    void clear_neighbors(Site& site, ArrayX<int16_t>& neighbor_count, int min_neighbors);
     /// Make an array of sublattice ids for the entire foundation
     ArrayX<sub_id> make_sublattice_ids(Foundation const& foundation);
 } // namespace detail
+
+/// Remove sites which have a neighbor count lower than `min_neighbors`
+void remove_dangling(Foundation& foundation, int min_neighbors);
 
 /**
  The foundation class creates a lattice-vector-aligned set of sites. The number of sites is high